% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrival.R
\name{expected_tau}
\alias{expected_tau}
\alias{expected_tau,hspec-method}
\title{Expected Inter-Arrival Time}
\usage{
expected_tau(
  object,
  rambda_component,
  type = 1,
  mu = NULL,
  beta = NULL,
  tol = .Machine$double.eps^0.25,
  max_upper = Inf,
  subdivisions = 400L
)
}
\arguments{
\item{object}{An object of class \code{hspec}.}

\item{rambda_component}{Rambda component.}

\item{type}{Process dimension index (default is 1).}

\item{mu}{Optional mu value (overrides object@mu if provided).}

\item{beta}{Optional beta value (overrides object@beta if provided).}

\item{tol}{Relative tolerance for numerical integration.}

\item{max_upper}{Upper integration limit.}

\item{subdivisions}{Number of subdivisions for numerical integration.}
}
\value{
Expected value of next inter-arrival time.
}
\description{
Computes the conditional expected time until the next event.
}
