% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmoment.R
\docType{methods}
\name{get_lambda0,hspec-method}
\alias{get_lambda0,hspec-method}
\title{Get the long-run expectation of intensities - lambda matrix}
\usage{
\S4method{get_lambda0}{hspec}(object)
}
\arguments{
\item{object}{hspec}
}
\description{
This function is not crucial but useful when you want to set lamba0 by default.
}
