% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsim.R
\docType{methods}
\name{hsim,hspec-method}
\alias{hsim,hspec-method}
\title{Simulate a (marked) Hawkes process}
\usage{
\S4method{hsim}{hspec}(object, lambda0 = NULL, n = 100)
}
\arguments{
\item{object}{\code{\link{hspec-class}}. This object includes the parameter values.}

\item{lambda0}{the starting values of lambda (intensity process). numeric or matrix.}

\item{n}{the number of observations.}
}
\description{
The method simulate marked Hawkes processes.
The object \code{\link{hspec-class}} contains the parameter values such as \code{mu}, \code{alpha}, \code{beta}.
The mark (jump) structure may or may not be included.
It returns an object of class '\code{\link{hreal-class}}.
}
