% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspec.R
\docType{class}
\name{hspec-class}
\alias{hspec-class}
\title{An S4 class to represent an exponential marked Hawkes model}
\description{
This class represents a specification of a marked Hawkes model with exponential kernel.
The intensity of the ground process is defined by:
\deqn{\lambda(t) = \mu + \int ( \alpha + \Psi ) * exp( -\beta (t-u)) d N(t)}
where \eqn{\alpha} represents exciting term that does not related to mark
\eqn{\Psi} represents the term related to mark and is represented by mark and impact.
}
\details{
\code{mu}, \code{alpha} and \code{beta} are required slots for every exponential Hawkes model.
\code{mark} and \code{impact} are additional slots.
}
\section{Slots}{

\describe{
\item{\code{mu}}{numeric value or matrix, automatically converted to matrix}

\item{\code{alpha}}{numeric value or matrix, automatically converted to matrix, exciting term}

\item{\code{beta}}{numeric value or matrix, automatically converted to matrix, exponential decay}

\item{\code{mark_hawkes}}{a function that represets mark for counting process}

\item{\code{mark_lambda}}{a function that represets mark for lambda process}

\item{\code{impact}}{a function that describe the after impact of mark to lambda}
}}

\examples{
MU <- matrix(c(0.2), nrow = 2)
ALPHA <- matrix(c(0.75, 0.92, 0.92, 0.75), nrow = 2, byrow=TRUE)
BETA <- matrix(c(2.25, 2.25, 2.25, 2.25), nrow = 2, byrow=TRUE)
mhspec2 <- new("hspec", mu=MU, alpha=ALPHA, beta=BETA)

}
