% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmoment.R
\name{hvol}
\alias{hvol}
\alias{hvol,hspec-method}
\title{Compute Hawkes volatility}
\usage{
hvol(
  object,
  horizon = 1,
  inter_arrival = NULL,
  type = NULL,
  mark = NULL,
  dependence = FALSE,
  lambda0 = NULL,
  ...
)

\S4method{hvol}{hspec}(
  object,
  horizon = 1,
  inter_arrival = NULL,
  type = NULL,
  mark = NULL,
  dependence = FALSE,
  lambda0 = NULL,
  ...
)
}
\arguments{
\item{object}{hspec}

\item{lambda0}{}
}
\description{
This function computes Hawkes volatilty.
}
