% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgfit.R
\docType{methods}
\name{infer_lambda,hspec-method}
\alias{infer_lambda,hspec-method}
\title{This method compute the infered lambda process and returns it as \code{hreal} form.
If we have realized path of Hawkes process and its parameter value, then we can compute the infered lambda processs.
Similarly with other method such as \code{hfit}, the input aruments are \code{inter_arrival}, \code{type}, \code{mark},
or equivalently, \code{N} and \code{Nc}.}
\usage{
\S4method{infer_lambda}{hspec}(object, inter_arrival = NULL, type = NULL,
  mark = NULL, N = NULL, Nc = NULL, lambda0 = NULL, N0 = NULL)
}
\description{
This method compute the infered lambda process and returns it as \code{hreal} form.
If we have realized path of Hawkes process and its parameter value, then we can compute the infered lambda processs.
Similarly with other method such as \code{hfit}, the input aruments are \code{inter_arrival}, \code{type}, \code{mark},
or equivalently, \code{N} and \code{Nc}.
}
