% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspec.R
\docType{methods}
\name{initialize,hspec-method}
\alias{initialize,hspec-method}
\title{Initialize the hspec object}
\usage{
\S4method{initialize}{hspec}(.Object, mu, alpha, beta, mark_hawkes = NULL,
  mark_lambda = NULL, impact = NULL, stability_check = FALSE)
}
\arguments{
\item{.Object}{hspec}

\item{mu}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{alpha}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{beta}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{mark_hawkes}{a function that generate mark}

\item{mark_lambda}{a function that represets mark for lambda process}

\item{impact}{a function that describe the mark impact}

\item{stability_check}{check the spectral radius}
}
\description{
\code{mu}, \code{alpha} and \code{beta} are required slots for every exponential Hawkes model.
\code{mark} and \code{impact} are additional slots.
}
