% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hllf.R
\docType{methods}
\name{logLik,hspec-method}
\alias{logLik,hspec-method}
\title{Compute the loglikelihood function}
\usage{
\S4method{logLik}{hspec}(object, inter_arrival, type = NULL, mark = NULL,
  N = NULL, Nc = NULL, N0 = NULL, lambda0 = NULL)
}
\arguments{
\item{object}{\code{\link{hspec-class}}. The parameter values in the object are used to compute the log-likelihood.}

\item{inter_arrival}{inter-arrival times of events. Includes inter-arrival for events that occur in all dimensions. Start with zero.}

\item{type}{a vector of dimensions. Distinguished by numbers, 1, 2, 3, and so on. Start with zero.}

\item{mark}{a vector of mark (jump) sizes. Start with zero.}

\item{N0}{the initial value of N}

\item{lambda0}{the initial values of lambda component. Must have the same dimensional matrix with \code{hspec}.}
}
\description{
The loglikelihood of the ground process of the Hawkes model.
(The estimation for jump distribution is not provided.)
}
\seealso{
\code{\link{hspec-class}}, \code{\link{hfit,hspec-method}}
}
