% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_unique_coef_mtrx}
\alias{name_unique_coef_mtrx}
\title{Unique naming coefficients from matrix}
\usage{
name_unique_coef_mtrx(M, notation)
}
\arguments{
\item{M}{a square matrix}

\item{notation}{a string for name}
}
\value{
to covert to matrix, use byrow=TRUE
}
\description{
With given matrix, look up the elements of the matrix
and assign the same name to the elements with the same value.
The name is based on \code{notation} and location of the element.
}
\examples{
A <- matrix(c(1,2,3,1), nrow = 2, byrow=TRUE)
name_unique_coef_mtrx(A, "alpha")

}
