% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmoment.R
\name{vol}
\alias{vol}
\alias{vol,hspec-method}
\title{Compute Hawkes volatility}
\usage{
vol(
  object,
  horizon = 1,
  inter_arrival = NULL,
  type = NULL,
  mark = NULL,
  dependence = FALSE,
  ...
)

\S4method{vol}{hspec}(
  object,
  horizon = 1,
  inter_arrival = NULL,
  type = NULL,
  mark = NULL,
  dependence = FALSE,
  ...
)
}
\arguments{
\item{object}{hspec}

\item{dependece}{}
}
\description{
Compute Hawkes volatility
}
