% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hreal.R
\name{hreal}
\alias{hreal}
\alias{print.hreal}
\alias{summary.hreal}
\alias{as.matrix.hreal}
\title{Generics for hreal}
\usage{
\method{print}{hreal}(x, n = 20, ...)

\method{summary}{hreal}(object, n = 20, ...)

\method{as.matrix}{hreal}(x, ...)
}
\arguments{
\item{x}{S3-object of hreal.}

\item{n}{number of rows to diplay.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{S3-object of hreal.}
}
\description{
Generic functions list for hreal:

Print the realization of the Haweks model.

Print the summary of the Hawkes process realization.

Matrix like ouput of the realization of Hawkes model.
}
