% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hreal.R
\name{hreal}
\alias{hreal}
\alias{print.hreal}
\alias{summary.hreal}
\alias{as.matrix.hreal}
\title{Generic for hreal}
\usage{
\method{print}{hreal}(res, n = 20, ...)

\method{summary}{hreal}(res, n = 20)

\method{as.matrix}{hreal}(res)
}
\arguments{
\item{res}{S3-object of hreal}

\item{n}{number of rows to diplay}
}
\description{
Generic functions list for hreal

Print the summary of the realization of the Haweks model.

This function presents the summary of the Hawkes realization.

The realization of Hawkes model is represented by matrix like ouput.
}
